fix: toObservableValue as projection
