+++ modules/component/src/core/index.ts
@@ -1,3 +1,3 @@
 export * from './utils';
-export * from './operators';
+export * from './projections';
 export * from './cd-aware.abstract';
