+++ modules/component/src/core/projections/index.ts
