+++ modules/component/src/core/projections/toObservableValue.ts
